Update buffer line style

+++ .config/nvim/lua/user/whichkey.lua
@@ -43,7 +43,7 @@ local setup = {
     scroll_up = '<C-u>', -- binding to scroll up inside the popup
   },
   window = {
-    border = 'double', -- none, single, double, shadow
+    border = 'single', -- none, single, double, shadow
     position = 'bottom', -- bottom, top
     margin = { 1, 0, 1, 0 }, -- extra window margin [top, right, bottom, left]
     padding = { 2, 2, 2, 2 }, -- extra window padding [top, right, bottom, left]
